@@ -1,38 +1,46 @@
 {
     "name": "jsonp-node",
-    "version": "0.1.1",
+    "version": "0.1.2",
     "description": "nodejs jsonp",
-    "main": "./bin/index.js",
-    "types": "./bin/index.d.ts",
-    "scripts": {
-        "test": "mocha --compilers ts:ts-node/register,tsx:ts-node/register ./test/index.test.ts",
-        "compile": "gulp compile"
-    },
     "keywords": [
         "jsonp",
         "parse",
         "stringify"
     ],
-    "author": "wujingtao",
+    "author": "wu-jingtao",
     "license": "MIT",
+    "homepage": "https://github.com/wu-jingtao/jsonp-node#readme",
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/mx601595686/jsonp-node.git"
+        "url": "git+https://github.com/wu-jingtao/jsonp-node.git"
     },
     "bugs": {
-        "url": "https://github.com/mx601595686/jsonp-node/issues"
+        "url": "https://github.com/wu-jingtao/jsonp-node/issues"
     },
-    "homepage": "https://github.com/mx601595686/jsonp-node#readme",
+    "main": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "lint": "eslint --max-warnings 0 \"{src,test}/**/*.{ts,mts,cts,tsx}\"",
+        "test": "mocha --require ts-node/register './test/**/*.test.ts'",
+        "compile": "tsc --project tsconfig.compile.json",
+        "compile-watch": "tsc --watch --project tsconfig.compile.json"
+    },
     "devDependencies": {
-        "@types/expect.js": "^0.3.29",
-        "@types/mocha": "^5.2.5",
-        "@types/node": "^10.7.1",
+        "@types/expect.js": "^0.3.32",
+        "@types/lodash": "^4.17.13",
+        "@types/mocha": "^10.0.9",
+        "@types/node": "^22.9.0",
+        "eslint-config-sufu": "^3.0.3",
         "expect.js": "^0.3.1",
-        "gulp": "^3.9.1",
-        "gulp-sourcemaps": "^2.6.4",
-        "gulp-typescript": "^5.0.0-alpha.3",
-        "mocha": "^5.2.0",
-        "ts-node": "^7.0.1",
-        "typescript": "^3.0.1"
+        "mocha": "^10.8.2",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.6.3"
+    },
+    "dependencies": {
+        "json5": "^2.2.3",
+        "tslib": "^2.8.1"
     }
 }
