@@ -1,70 +1,50 @@
-import * as vm from 'vm';
-
-//#region 解析标准jsonp格式
+import json5 from 'json5';
 
 /**
- * 解析jsonp
- * @param jsonp jsonp数据
- * @param callback jsonp指定的回调函数名称
+ * 解析 jsonp
+ * @param data jsonp 数据
  */
-function parse(jsonp: string, callback: string) {
-    const ctx = {
-        result: null as any,
-        [callback]: function (data: any) {
-            ctx.result = data;
-        }
-    };
-
-    vm.runInNewContext(jsonp, ctx);
-
-    return ctx.result;
+function parse(data: string): unknown {
+    const start = data.indexOf('(') + 1;
+    const end = data.lastIndexOf(')');
+    data = data.slice(start, end);
+    return json5.parse(data);
 }
 
 /**
- * 将数据序列化成jsonp格式
+ * 将数据序列化成 jsonp 格式
  * @param data 要序列化的数据
- * @param callback 回调函数的名称，默认为随机数
+ * @param callbackName 回调函数的名称
  */
-function stringify(data: any, callback: string) {
-    return `${callback}(${JSON.stringify(data)})`;
+function stringify(data: any, callbackName?: string): string {
+    return `${callbackName}(${JSON.stringify(data)});`;
 }
 
-//#endregion
-
-//#region 解析 变量格式
-
 /**
- * 解析变量格式的jsonp
- * @param jsonp jsonp数据
- * @param varName 变量名称
- * @param isGlobal 是否是全局变量
+ * 解析变量格式的 jsonp
+ * @param data jsonp 数据
  */
-function parse_var(jsonp: string, varName: string, isGlobal: boolean) {
-    if (isGlobal) {
-        const ctx = { [varName]: null as any };
+function parse_var(data: string): unknown {
+    const matchStart = data.match(/=\s*/);
+    const matchEnd = data.match(/\s*;?\s*$/);
 
-        vm.runInNewContext(jsonp, ctx);
-
-        return ctx[varName];
+    if (matchStart === null) {
+        throw new Error('Data format error.');
     } else {
-        const ctx = { result: null as any };
-
-        vm.runInNewContext(`${jsonp}; result = ${varName}`, ctx);
-
-        return ctx.result;
+        const start = matchStart.index! + matchStart[0].length;
+        const end = matchEnd!.index;
+        return json5.parse(data.slice(start, end));
     }
 }
 
 /**
- * 将数据序列化成变量jsonp格式
+ * 将数据序列化成 jsonp 变量格式
  * @param data 要序列化的数据
  * @param varName 变量名称
- * @param isGlobal 是否是全局变量
+ * @param isGlobal 是否是全局变量(在变量名前面添加 var 关键字)
  */
-function stringify_var(data: any, varName: string, isGlobal: boolean) {
-    return `${isGlobal ? '' : 'var '}${varName}=${JSON.stringify(data)}`;
+function stringify_var(data: any, varName: string, isGlobal: boolean): string {
+    return `${isGlobal ? '' : 'var '}${varName}=${JSON.stringify(data)};`;
 }
-
-//#endregion
 
 export = { parse, stringify, parse_var, stringify_var };
